chore(lint): fix clippy warnings and format code

+++ src/stages/map/rewrite.rs
@@ -123,7 +123,7 @@ fn normalize_lut_cell_inputs(cell: &mut Cell) -> bool {
         let mut old_address = 0usize;
         for (old_index, maybe_new_index) in old_to_new.iter().enumerate() {
             let bit = maybe_new_index
-                .map(|new_index| ((new_address >> new_index) & 1) as usize)
+                .map(|new_index| (new_address >> new_index) & 1)
                 .unwrap_or(0);
             old_address |= bit << old_index;
         }
